+++ src/main/java/com/assetflow/reservation/dto/ReservationCreateRequest.java
@@ -1,16 +1,15 @@
 package com.assetflow.reservation.dto;
 
-
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
-@AllArgsConstructor
+@NoArgsConstructor
 public class ReservationCreateRequest {
-        @NotBlank
+        @NotNull
         private Long memberId;
-        @NotBlank
+        @NotNull
         private Long assetItemId;
 }